chore: gen fresh bolt11s per melt, drop sidecar
The `bolt11-gen` one-shot sidecar pre-baked a fixed invoice set, causing mint rejections on replay (duplicate `payment_hash`). Replace with per-call generation: CLN stacks call `lightning-cli invoice` via `LN_INVOICE_NODE`; fake stacks self-sign via `gen-one-bolt11.js` bundled in a new `activity-fake.Dockerfile`.
