+++ e2e/docker/scripts/activity-fake.sh
@@ -3,9 +3,12 @@
 #
 # Runs once from the `activity` compose service after orchard is healthy.
 # Exercises mint → swap → melt flows via cdk-cli for each configured unit.
-# No LN node is touched — the mint's fake backend (cdk fake_wallet or
-# nutshell FakeWallet) auto-confirms quotes, and melts point at bolt11
-# fixtures pre-generated by the bolt11-gen sidecar.
+# Melts need a fresh bolt11 per call (the mint rejects re-melts of the same
+# payment_hash with code 11000 "already paid or pending"). Two paths:
+#   - LN_INVOICE_NODE set → `docker exec <node> lightning-cli invoice ...`
+#     against the existing CLN node. Used on cln-nutshell-postgres.
+#   - else → `node /scripts/gen-one-bolt11.js <sats>` self-signs locally.
+#     Used on fake-cdk-postgres which has no LN node.
 #
 # Designed to run on:
 #   - fake-cdk-postgres (sat + usd both via fake)
@@ -20,9 +23,10 @@
 #   - ACTIVITY_MEMPOOL_PER_RATE
 #
 # Requires:
-#   - docker socket mounted (for exec into wallet)
-#   - /shared mounted ro (fake-bolt11s.json)
+#   - docker socket mounted (for exec into wallet and optionally LN node)
 #   - CONFIG_NAME, MINT_SERVICE, MINT_PORT env vars set by compose
+#   - LN_INVOICE_NODE (optional): short name of a CLN container to ask
+#     for melt invoices, e.g. "cln-alice". Resolved as ${CONFIG_NAME}-${LN_INVOICE_NODE}.
 
 set -eu
 
@@ -31,7 +35,6 @@ set -eu
 : "${MINT_PORT:?MINT_PORT must be set (3341|3340)}"
 
 MINT_URL="http://${MINT_SERVICE}:${MINT_PORT}"
-FIXTURES="/shared/fake-bolt11s.json"
 
 ACTIVITY_MINTS_SAT=${ACTIVITY_MINTS_SAT:-5}
 ACTIVITY_MINTS_USD=${ACTIVITY_MINTS_USD:-3}
@@ -51,11 +54,6 @@ if [ "${ACTIVITY_SKIP:-0}" = "1" ]; then
     exit 0
 fi
 
-if [ ! -f "$FIXTURES" ]; then
-    log "fixtures file not found: $FIXTURES (bolt11-gen sidecar should have written it)"
-    exit 1
-fi
-
 # bitcoind JSON-RPC wrapper (mirrors activity-cln-topology.sh).
 bcli() {
     method="$1"; params="${2:-[]}"
@@ -73,30 +71,33 @@ wallet_unit() {
     docker exec -i "${CONFIG_NAME}-wallet" cdk-cli --unit "$unit" "$@"
 }
 
-# Fixture amounts — keep in sync with gen-bolt11s.js:AMOUNTS.
-FIXTURE_AMOUNTS="100 500 1000"
-
-# Load every (unit × amount) fixture once into shell vars (FIXTURE_<unit>_<amt>);
-# melt loops pick from memory instead of spawning jq per iteration.
-for unit in sat usd eur; do
-    for amt in $FIXTURE_AMOUNTS; do
-        eval "FIXTURE_${unit}_${amt}=$(jq -r ".${unit}.\"${amt}\"" "$FIXTURES")"
-    done
-done
+AMOUNTS="100 500 1000"
 
 rand_amount() {
     r=$(od -An -N1 -tu1 < /dev/urandom | tr -cd '0-9')
     idx=$((r % 3))
     i=0
-    for a in $FIXTURE_AMOUNTS; do
+    for a in $AMOUNTS; do
         if [ "$i" -eq "$idx" ]; then printf '%s' "$a"; return; fi
         i=$((i + 1))
     done
 }
 
-fixture_for() {
-    unit="$1"; amt="$2"
-    eval "printf '%s' \"\${FIXTURE_${unit}_${amt}:-}\""
+# Echo a fresh bolt11 invoice for $amt sats. CLN path uses a unique label per
+# call so the node never rejects a duplicate; local path uses random privkey +
+# payment_hash per invocation so the mint sees a fresh payment_hash every time.
+fresh_bolt11() {
+    amt_sat="$1"
+    if [ -n "${LN_INVOICE_NODE:-}" ]; then
+        msat=$((amt_sat * 1000))
+        rid=$(od -An -N4 -tu4 < /dev/urandom | tr -cd '0-9')
+        docker exec -i "${CONFIG_NAME}-${LN_INVOICE_NODE}" \
+            lightning-cli --lightning-dir=/home/clightning/.lightning --network=regtest \
+            invoice "$msat" "activity-fake-${rid}" "e2e-fake-melt" 2>/dev/null \
+            | jq -r '.bolt11'
+    else
+        node /scripts/gen-one-bolt11.js "$amt_sat"
+    fi
 }
 
 # ── Per-unit mint/swap/melt loop ──
@@ -142,9 +143,9 @@ run_unit() {
         i=0
         while [ "$i" -lt "$n_melts" ]; do
             amt=$(rand_amount)
-            inv=$(fixture_for "$unit" "$amt")
+            inv=$(fresh_bolt11 "$amt")
             if [ -z "$inv" ] || [ "$inv" = "null" ]; then
-                log "  [$unit] melt ${amt} FAILED (no fixture)"
+                log "  [$unit] melt ${amt} FAILED (invoice gen)"
                 i=$((i + 1)); continue
             fi
             if wallet_unit "$unit" melt --mint-url "$MINT_URL" --invoice "$inv" >/dev/null 2>&1; then
