+++ package.json
@@ -38,8 +38,8 @@
     "e2e:down": "./e2e/docker/scripts/compose.sh down",
     "e2e:logs": "./e2e/docker/scripts/compose.sh logs",
     "e2e:ps": "./e2e/docker/scripts/compose.sh ps",
-    "e2e:sim:start": "./e2e/docker/scripts/compose.sh activity start",
-    "e2e:sim:stop": "./e2e/docker/scripts/compose.sh activity stop",
+    "e2e:sim:up": "./e2e/docker/scripts/compose.sh activity start",
+    "e2e:sim:down": "./e2e/docker/scripts/compose.sh activity stop",
     "e2e:sim:status": "./e2e/docker/scripts/compose.sh activity status",
     "e2e:sim:logs": "./e2e/docker/scripts/compose.sh activity logs",
     "e2e:bootstrap-mainchain": "./e2e/docker/scripts/bootstrap-mainchain.sh",
